+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 1.0 (unreleased)
 ----------------
 
+- Assign a portlet onto the Pluggable Login page on installation.
+  [davidjb]
 - Change from a Embedded WAYF to Shibboleth's EDS.
   [davidjb]
 - Remove old ``collective.aaf`` code from this package; that package now